is word guessed iterates over each char in secret_word and appends it a list secret_word_char_list.
next for every char in secret_word if that char is in the list
letters_guessed that char is appended to valid_guesses list.

If valid_guesses and secret_word_char_list are equal function returns
valid. Else function returns false.

I want to use this as a test to end the game. I may also be able to
use these list as basis for the game scoring objective in the pset pdf.